adjusted crud and router functions

+++ app/routers/tune_router.py
@@ -13,7 +13,7 @@ router = APIRouter(
     prefix="/tunes"
 )
 
-IMAGE_DIR = "app/uploads/"
+UPLOAD_DIR = "app/uploads/"
 
 
 @router.post("/", response_model=Tune)
@@ -24,34 +24,45 @@ async def create_tune(
     file: UploadFile = File(...),
     user_id: int = Form(...),
     db: Session = Depends(dependencies.get_db)
-):
-    base_url = str(request.base_url)  
+): 
     contents = await file.read()
-    image_url = f"{base_url}uploads/{file.filename}"
-    tune_in_db=crud.get_tune(db=db, code=code)
+    file_path = os.path.join(UPLOAD_DIR, file.filename)
+    if os.path.exists(file_path):
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=f"File with the name {file.filename} exits")
+    tune_in_db = crud.get_tune(db=db, code=code)
     if tune_in_db:
-        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Tune already exist!")
-    tune_data = TuneCreate(name=name, code=code, user_id=user_id, file_name=image_url)
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=f"Tune with code {code} already exist!")
+    
+    tune_data =TuneCreate(name=name, code=code, user_id=user_id, file_name=file.filename)
     db_tune = crud.create_tune(db=db, tune=tune_data)
-    file_path = f"app/uploads/{file.filename}"
-    file_path = os.path.join(IMAGE_DIR,file.filename)
+    
     with open(file_path, "wb") as f:
         f.write(contents)
+
     return db_tune
 
 @router.get("/")
-async def read_tunes(skip: int = 0, limit: int = 100, searchText:Optional[str]='', 
+async def read_tunes(request: Request, skip: int = 0, limit: int = 100, searchText:Optional[str]='', 
                      db: Session = Depends(dependencies.get_db)):
     tunes = crud.get_tunes(db, skip=skip, limit=limit, search=searchText)
     if not tunes:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No SKiza tunes found")
+    for tune in tunes:
+        tune.file_name = f"{request.url.scheme}://{request.url.netloc}/uploads/{tune.file_name}"
     return tunes
 
 @router.delete("/{tune_id}")
 def delete_tune(tune_id: int,  db: Session = Depends(dependencies.get_db)):
-    db_tune = crud.delete_tune(db=db, tune_id=tune_id)
+    db_tune = crud.get_tune_by_id(db=db, tune_id=tune_id)
     if db_tune is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail={"message":"tune not found!"})
+    
+    file_path = os.path.join("app/uploads", db_tune.file_name)
+
+    if os.path.exists(file_path):
+        os.remove(file_path)
+    
+    crud.delete_tune(db=db, tune_id=tune_id)
     return db_tune
 
 @router.put("/{tune_id}", response_model=Tune)
@@ -68,7 +79,7 @@ async def edit_tune(
     cleaned_name = name.replace(" ", "")
     image_url = f"{base_url}uploads/{cleaned_name}"
     contents = await file.read()
-    file_path = os.path.join(IMAGE_DIR, cleaned_name)
+    file_path = os.path.join(UPLOAD_DIR, cleaned_name)
 
     with open(file_path, "wb") as f:
         f.write(contents)
